requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# TradeRecon AI Agent — Python Dependencies
# ══════════════════════════════════════════════════════════════════════════════
# Install: pip install -r requirements.txt
# Upgrade:  pip install -r requirements.txt --upgrade
# ══════════════════════════════════════════════════════════════════════════════

# ── AI / LLM ──────────────────────────────────────────────────────────────────
anthropic>=0.40.0

# ── Data Wrangling ────────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.2
xlrd>=2.0.1

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Configuration & Environment ───────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── Scheduling ────────────────────────────────────────────────────────────────
APScheduler>=3.10.4

# ── Templating (HTML email reports) ──────────────────────────────────────────
jinja2>=3.1.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ── Code Quality ──────────────────────────────────────────────────────────────
ruff>=0.1.0
black>=23.12.0

# ── Utilities ─────────────────────────────────────────────────────────────────
rich>=13.7.0
python-dateutil>=2.8.2
